split_registry_snapshots
After discussion we decided to rename this endpoint and fold the build_timestamp validation logic into SplitRegistry. Besides making this more of a true resource route, it explains what this endpoint is retunring much better, and consolidates code. Now the SplitRegistrySnapshot class accepts a timestamp and returns to you the split registry as of that time. So this change will be backwards compatifble as we just need to pass it `Time.zone.now`
