--- app/controllers/api/v2/legacy_split_registries_controller.rb
@@ -1,7 +1,0 @@
-class Api::V2::LegacySplitRegistriesController < UnauthenticatedApiController
-  include CorsSupport
-
-  def show
-    @split_registry = SplitRegistry.new(Time.zone.now)
-  end
-end
