+++ app/controllers/api/v2/split_registries_controller.rb
@@ -2,17 +2,6 @@ class Api::V2::SplitRegistriesController < UnauthenticatedApiController
   include CorsSupport
 
   def show
-    build_timestamp = BuildTimestamp.new(build_params)
-    if build_timestamp.valid?
-      @split_registry = SplitRegistry.new(build_timestamp.build_timestamp)
-    else
-      render_errors build_timestamp
-    end
-  end
-
-  private
-
-  def build_params
-    params.permit(:timestamp)
+    @split_registry = SplitRegistrySnapshot.new(timestamp: Time.zone.now)
   end
 end
