+++ app/controllers/api/v2/split_registry_snapshots_controller.rb
@@ -1,0 +1,19 @@
+class Api::V2::SplitRegistrySnapshotsController < UnauthenticatedApiController
+  include CorsSupport
+
+  def show
+    snapshot = SplitRegistrySnapshot.new(timestamp: snapshot_params[:timestamp])
+
+    if snapshot.valid?
+      @split_registry_snapshot = snapshot
+    else
+      render_errors snapshot
+    end
+  end
+
+  private
+
+  def snapshot_params
+    params.permit(:timestamp)
+  end
+end
